Expose `DEEPEN_LENGTH` through the `deepen_length` action input
Consumers of this action may wish to customize the `--deepen` parameter
used while fetching more commits. This is useful because they may have
an average PR commit length that's much smaller or much larger than the
default of 10, so tuning this parameter can optimize the `git fetch`
calls for their repository. (The tradeoff of setting a large number is
that we may fetch more unnecessary commits when running the action on a
PR that only has a few commits. On the other hand, setting a small
number may lead to many `git fetch` calls in a row in order to fetch all
the commits of a large PR, with each call incurring some overhead.)

It was already possible to customize this parameter through the
`DEEPEN_LENGTH` environment variable, but exposing it through an action
input is a bit more ergonomic (especially since we already offer two
other inputs).